requirements.txt
fastapi[all]
python-dotenv
tiktoken
pydantic